fix: refactor job status
