+++ Command/AddRecurringConsoleJobToQueueCommand.php
@@ -2,7 +2,7 @@
 
 namespace Markup\JobQueueBundle\Command;
 
-use Markup\JobQueueBundle\Factory\JobStatusFactory;
+use Markup\JobQueueBundle\Repository\JobStatusRepository;
 use Markup\JobQueueBundle\Model\RecurringConsoleCommandConfiguration;
 use Markup\JobQueueBundle\Repository\CronHealthRepository;
 use Markup\JobQueueBundle\Repository\JobLogRepository;
@@ -48,22 +48,22 @@ class AddRecurringConsoleJobToQueueCommand extends Command
      */
     private $environment;
 
-    /** @var JobStatusFactory */
-    private $jobStatusFactory;
+    /** @var JobStatusRepository */
+    private $jobStatusRepository;
 
     public function __construct(
         RecurringConsoleCommandReader $recurringConsoleCommandReader,
         JobManager $jobManager,
         CronHealthRepository $cronHealthRepository,
         JobLogRepository $jobLogRepository,
-        JobStatusFactory $jobStatusFactory,
+        JobStatusRepository $jobStatusRepository,
         string $environment
     ) {
         $this->recurringConsoleCommandReader = $recurringConsoleCommandReader;
         $this->jobManager = $jobManager;
         $this->cronHealthRepository = $cronHealthRepository;
         $this->jobLogRepository = $jobLogRepository;
-        $this->jobStatusFactory = $jobStatusFactory;
+        $this->jobStatusRepository = $jobStatusRepository;
         $this->environment = $environment;
 
         parent::__construct(null);
@@ -133,8 +133,12 @@ class AddRecurringConsoleJobToQueueCommand extends Command
                 }
             }
 
-            if ($configuration->getUserManaged() && !$this->jobStatusFactory->isUserEnabledJob($configuration->getCommand(), $configuration->getArguments())) {
-                continue;
+            if ($configuration->isUserManaged()) {
+                $arguments = $configuration->getArguments() ? json_encode($configuration->getArguments()) : null;
+
+                if (!$this->jobStatusRepository->isStatusEnabled($command, $arguments)) {
+                    continue;
+                }
             }
 
             $this->jobManager->addConsoleCommandJob(
