+++ Service/RecurringConsoleCommandReader.php
@@ -6,7 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Markup\JobQueueBundle\Exception\InvalidConfigurationException;
 use Markup\JobQueueBundle\Exception\MissingScheduleException;
 use Markup\JobQueueBundle\Exception\MissingConfigurationException;
-use Markup\JobQueueBundle\Factory\JobStatusFactory;
+use Markup\JobQueueBundle\Repository\JobStatusRepository;
 use Markup\JobQueueBundle\Model\RecurringConsoleCommandConfiguration;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
@@ -38,15 +38,15 @@ class RecurringConsoleCommandReader
      * @param string $kernelPath
      */
 
-    /** @var JobStatusFactory */
-    private $jobStatusFactory;
+    /** @var JobStatusRepository */
+    private $jobStatusRepository;
 
     public function __construct(
         string $kernelPath,
-        JobStatusFactory $jobStatusFactory
+        JobStatusRepository $jobStatusRepository
     ) {
         $this->kernelPath = $kernelPath;
-        $this->jobStatusFactory = $jobStatusFactory;
+        $this->jobStatusRepository = $jobStatusRepository;
     }
 
     public function setConfigurationFileName($name)
@@ -149,13 +149,20 @@ class RecurringConsoleCommandReader
                 isset($group['timeout']) ? $group['timeout'] : null,
                 isset($group['envs']) ? $group['envs'] : null,
                 isset($group['user_managed']) ? $group['user_managed'] : null,
-                isset($group['user_managed']) ? $this->jobStatusFactory->isStatusEnabled($group['command'], isset($group['arguments']) ? json_encode($group['arguments']) : null) : null
+                isset($group['user_managed']) ? $this->jobStatusEnabled($group) : null
             );
 
             $configurations->add($recurringConsoleCommandConfiguration);
         }
 
         return $configurations;
+    }
+
+    private function jobStatusEnabled(array $group): bool
+    {
+        $arguments = isset($group['arguments']) ? json_encode($group['arguments']) : null;
+
+        return $this->jobStatusRepository->isStatusEnabled($group['command'], $arguments);
     }
 
     /**
